backend/requirements.txt
# requirements.txt

django==4.2.14
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==2.3.0
django-cors-headers==4.3.1